fix(communication): send inspection protocol emails via Infobip SDK
